requirements.txt
ultralytics==8.0.145
streamlit==1.25.0
pafy==0.5.5
youtube-dl==2021.12.17
opencv-python~=4.8.0.74
torch~=2.0.1
numpy~=1.25.1
ppq~=0.6.6
tqdm~=4.65.0
onnxruntime~=1.16.3
onnx~=1.16.0
torchvision~=0.15.2
pillow~=9.5.0
pandas~=2.0.3
setuptools~=68.0.0
mmdeploy~=0.2.0
mmengine~=0.10.4
packaging~=23.1
mmcv~=2.1.0
pyyaml~=6.0.1
h5py~=3.11.0
mmdet~=3.3.0